// lib/domain/main.types.ts
export interface TimeStamps {
    createdAt: number
    updatedAt: number
}

export type AcceptablePKValue =
    | `${AcceptablePK.YOUTUBE_VIDEO}/${string}`
    | `${AcceptablePK.YOUTUBE_NOTIFICATION}/${string}`
    | `${AcceptablePK.TRANSCRIPTION}/${string}`
    | `${AcceptablePK.SUMMARY}/${string}`
    | AcceptablePK.SUBSCRIBED_CHANNEL
    | AcceptablePK.TELEGRAM_CHANNEL

export interface MainItem extends TimeStamps {
    [MainTable.PK]: AcceptablePKValue
    [MainTable.SK]: string
    expireAt?: number
}
// ...
export interface VideoSummaryItem<T> extends MainItem {
    [MainTable.PK]: `${AcceptablePK.SUMMARY}/${string}`
    [MainTable.SK]: UUID
    [MainTable.TIMESTAMP]: number
    videoId: string
    videoTitle: string
    [VIDEO_TYPE_KEY]: YoutubeVideoType
    [VIDEO_GENRE_KEY]: Exclude<VideoGenre, 'ALARM'>
    channelTitle: string
    channelUri: string
    summary: T
}
// ...
export type TinfoilSummaryResults = Readonly<{
    summary: string[]
    shortSummary: string
    absurdityLevel: number
}>

export type SoftwareEngineeringSummaryResults = Readonly<{
    summary: string[]
    shortSummary: string
    topics: Array<string>
    technologies: Array<string>
    codePresent: boolean
    languagesMentioned: Array<string>
    difficulty: string
    audience: string
}>

export type PoliticsSummaryResults = Readonly<{
    summary: string[]
    shortSummary: string
    keyActors: string[]
    mainIssues: string[]
    tone: string
    impactAssessment: string
}>

export type ScienceSummaryResults = Readonly<{
    summary: string[]
    shortSummary: string
    keywords: string[]
    complexityLevel: string
}>

export type AcceptableLlmResponse =
    | TinfoilSummaryResults
    | SoftwareEngineeringSummaryResults
    | PoliticsSummaryResults
    | ScienceSummaryResults

export type TranscriptVideo = Readonly<{
    id: string
    channelId: string
    videoId: string
    videoTitle: string
    [VIDEO_TYPE_KEY]: YoutubeVideoType
    [VIDEO_GENRE_KEY]: Exclude<VideoGenre, 'ALARM'>
    channelTitle: string
    channelUri: string
    publishedAt: number
    transcript: string
}> &
    Message

export type YoutubeVideoType = 'SHORT' | 'STANDARD' | 'LIVE' | 'UNKNOWN' | 'UPCOMING' | 'LONG'

export type VideoGenre = 'TINFOIL' | 'SOFTWARE_ENGINEERING' | 'ALARM' | 'POLITICS' | 'SCIENCE'

export type ErrorOutput = Readonly<{
    [ERROR_OUTPUT_ATTR_KEY]: string
    payload: object
    subject: string
}>

export type ChatNewsMessageInputPayload = Readonly<{
    [VIDEO_GENRE_KEY]: VideoGenre
    message: string
}>

export type Message = Readonly<{
    type: MessageType
    createdAt: number
    sendAt: number
}>